requirements.txt
gym
numpy
treys
pyyaml
stable-baselines3[extra]